+++ src/conversions/audio_to_stereo.ml
@@ -41,7 +41,9 @@ class basic ~kind source =
           (* Set audio layer. *)
           let audio =
             match Frame.(tmp_frame.content.audio) with
-              | [||] -> assert false
+              | [||] ->
+                  let buf = Audio.Mono.create (AFrame.size ()) in
+                  [| buf; buf |]
               | [| chan |] -> [| chan; chan |]
               | audio -> Array.sub audio 0 2
           in
@@ -49,10 +51,13 @@ class basic ~kind source =
   end
 
 let () =
-  let input_kind = Lang.kind_type_of_kind_format Lang.audio_any in
+  let input_kind = Lang.kind_type_of_kind_format Lang.any in
+  let { Frame.audio; video; midi } = Lang.of_frame_kind_t input_kind in
+  ignore audio;
+  let output_kind = Lang.frame_kind_t ~audio:(Lang.n_t 2) ~video ~midi in
   Lang.add_operator "audio_to_stereo" ~category:Lang.Conversions
     ~descr:"Convert any kind of audio source into a stereo source."
-    ~return_t:(Lang.kind_type_of_kind_format Lang.audio_stereo)
+    ~return_t:output_kind
     [("", Lang.source_t input_kind, None, None)]
     (fun p kind ->
       let s = new basic ~kind (Lang.to_source (List.assoc "" p)) in